requirements.txt
blinker==1.6.2
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==42.0.7
Deprecated==1.2.14
Flask==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.3
packaging==23.1
pycparser==2.21
PyGithub==1.59.0
PyJWT==2.8.0
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.15.0
